feat: fix bugs
