+++ app/(news-guides)/guide/page.tsx
@@ -1,5 +1,6 @@
 import NewbieData from '@/components/newbie-guide/NewbieData';
 import HeadingSection from '@/components/reusable/HeadingSection';
+import { guideData } from '@/lib/dataSample';
 
 export const metadata = {
 	title: 'Ran online Elitegrounds | Guide'
@@ -11,7 +12,7 @@ export default function page() {
 			<HeadingSection title='Guide' />
 
 			<section className='max-w-5xl mx-auto my-5 lg:my-10 grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 justify-items-center gap-y-10'>
-				<NewbieData />
+				<NewbieData dataArr={guideData} />
 			</section>
 		</section>
 	);
