+++ app/(news-guides)/news/page.tsx
@@ -1,6 +1,8 @@
 import NewbieData from '@/components/newbie-guide/NewbieData';
 import HeadingSection from '@/components/reusable/HeadingSection';
 
+import { newsData } from '@/lib/dataSample';
+
 export const metadata = {
 	title: 'Ran online Elitegrounds | News'
 };
@@ -11,7 +13,7 @@ export default function page() {
 			<HeadingSection title='News' />
 
 			<section className='max-w-5xl mx-auto my-5 lg:my-10 grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 justify-items-center gap-y-10'>
-				<NewbieData />
+				<NewbieData dataArr={newsData} />
 			</section>
 		</section>
 	);
